requirements.txt
aiosqlite
discord.py
git+https://github.com/bijij/Flag-Parsing
jishaku
git+https://github.com/jucacrispim/aiomediawiki
